# docs/swagger.yaml
basePath: /v1
definitions:
...
  enum.DestinationStatus:
    enum:
    - 0
    - 1
    - 2
    type: integer
    x-enum-varnames:
    - Editing
    - Confirm
    - Deleted
  model.Compound:
    properties:
      commune:
        type: string
      district:
        type: string
      province:
        type: string
    type: object
  model.Destination:
    properties:
      arriveTime:
        type: string
      createdTime:
        type: string
      id:
        format: uuid
        type: string
      lat:
        type: number
      lng:
        type: number
      location:
        $ref: '#/definitions/model.Location'
      locationId:
        type: string
      name:
        type: string
      status:
        $ref: '#/definitions/enum.DestinationStatus'
      stayOverTime:
        description: đơn vị là phút
        type: integer
      updatedTime:
        type: string
    type: object
...
  model.Location:
    properties:
      address:
        type: string
      compound:
        $ref: '#/definitions/model.Compound'
      createdTime:
        type: string
      formattedAddress:
        type: string
      id:
        format: uuid
        type: string
      isVerified:
        type: boolean
      lat:
        type: number
      lng:
        type: number
      name:
        type: string
      placeId:
        type: string
      placeType:
        $ref: '#/definitions/model.PlaceType'
      placeTypeId:
        type: string
      updatedTime:
        type: string
    type: object
  model.PlaceType:
    properties:
      createdTime:
        type: string
      id:
        format: uuid
        type: string
      typeName:
        type: string
      updatedTime:
        type: string
    type: object
...
  request.UpdateDestinationRequest:
    properties:
      arriveTime:
        type: string
      lat:
        type: number
      lng:
        type: number
      name:
        type: string
      status:
        $ref: '#/definitions/enum.DestinationStatus'
      stayOverTime:
        type: integer
    type: object
...
  request.UpdateTripMemberRequest:
    properties:
      nickname:
        maxLength: 64
        minLength: 1
        type: string
      role:
        example: admin
        type: string
    type: object
...
  response.TripMemberResponse:
    properties:
      invitorName:
        type: string
      joinedAt:
        type: string
      nickname:
        type: string
      role:
        example: member
        type: string
      status:
        example: active
        type: string
      tripId:
        format: uuid
        type: string
      tripName:
        type: string
      userId:
        format: uuid
        type: string
      username:
        type: string
    type: object
  share.ErrorResponse:
    properties:
      errors:
        additionalProperties:
          type: string
        type: object
      message:
        example: invalid input
        type: string
      status:
        example: 400
        type: integer
    type: object
info:
  contact: {}
  description: 'Backend API for Road2D (team Vandra): collaborative multi-branch ride
    planning and group-ride safety in Vietnam. JWT auth, Goong Maps playground, trip
    graph models.'
  title: Road2D
  version: "1.0"
paths:
...
  /planning/destination/{destinationId}:
    delete:
      description: Deletes a destination that is still editing. A pin that sits only
        on draft branches is removed immediately. A pin on a route branch is stripped
        first; the delete is rejected when that would disconnect the graph or when
        a trip that is no longer planning still uses the pin on a route. Travels that
        pointed at the pin are removed with the row.
      parameters:
      - description: Destination UUID
        format: uuid
        in: path
        name: destinationId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: Destination deleted
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/share.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/share.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/share.ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/share.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Delete destination
      tags:
      - planning
    get:
      description: Returns a destination. When locationId is set, lat/lng are locked
        to that location.
      parameters:
      - description: Destination UUID
        format: uuid
        in: path
        name: destinationId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Destination'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/share.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/share.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/share.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get destination
      tags:
      - planning
    put:
      consumes:
      - application/json
      description: Updates name, arrive time, stay, and status while the destination
        is editing. lat/lng are applied only when locationId is empty; a forked or
        verified pin keeps the linked location's coordinates.
      parameters:
      - description: Destination UUID
        format: uuid
        in: path
        name: destinationId
        required: true
        type: string
      - description: Fields to store
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/request.UpdateDestinationRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Destination'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/share.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/share.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/share.ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/share.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update destination
      tags:
      - planning
...
  /trips/{tripId}/members/{userId}:
    delete:
      description: Leader or admin. Cannot kick yourself, the leader, or (as an admin)
        another admin. The row stays as kicked so the user can be invited again.
      parameters:
      - description: Trip UUID
        format: uuid
        in: path
        name: tripId
        required: true
        type: string
      - description: Member UUID
        format: uuid
        in: path
        name: userId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: Member kicked
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/share.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/share.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/share.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Kick trip member
      tags:
      - trips
    patch:
      consumes:
      - application/json
      description: An active member may change only their own nickname. The leader
        may also set another member's role to admin or member. Leadership is not assigned
        here.
      parameters:
      - description: Trip UUID
        format: uuid
        in: path
        name: tripId
        required: true
        type: string
      - description: Member UUID
        format: uuid
        in: path
        name: userId
        required: true
        type: string
      - description: Nickname and/or role
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/request.UpdateTripMemberRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.TripMemberResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/share.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/share.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/share.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/share.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update trip member
      tags:
      - trips
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
  PlaygroundKey:
    description: Shared secret from GOONG_PLAYGROUND_KEY. Scalar → Authorize.
    in: header
    name: X-Playground-Key
    type: apiKey
swagger: "2.0"